app/requirements.txt
requests
lxml
python-dotenv
bs4
pandas
langchain
langchain_core
langchain_ollama
openpyxl
langchain-openai
gspread
gspread_dataframe
oauth2client
